Moved tabs logic tabs component.
