+++ src/Components/Results/Results.tsx
@@ -9,12 +9,8 @@ import { Navigate, useParams } from 'react-router-dom';
 import ResultsHeader from './Header/Header';
 import ProgramPage from './ProgramPage/ProgramPage';
 import ResultsTabs from './Tabs/Tabs';
-import Needs from './Needs/Needs';
-import Programs from './Programs/Programs';
 import MoreHelp from './MoreHelp/MoreHelp';
 import NavigatorPage from './NavigatorPage/NavigatorPage';
-
-import { Box, Container, Tab, Tabs, Typography } from '@mui/material';
 
 type WrapperResultsContext = {
   programs: Program[];
@@ -26,38 +22,6 @@ type WrapperResultsContext = {
 type ResultsProps = {
   type: 'program' | 'need' | 'navigator';
 };
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: number;
-  value: number;
-}
-
-function CustomTabPanel(props: TabPanelProps) {
-  const { children, value, index, ...other } = props;
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`simple-tabpanel-${index}`}
-      aria-labelledby={`simple-tab-${index}`}
-      {...other}
-    >
-      {value === index && (
-        <Box sx={{ p: 3 }}>
-          <Typography>{children}</Typography>
-        </Box>
-      )}
-    </div>
-  );
-}
-
-function a11yProps(index: number) {
-  return {
-    id: `simple-tab-${index}`,
-    'aria-controls': `simple-tabpanel-${index}`,
-  };
-}
 
 export const ResultsContext = createContext<WrapperResultsContext | undefined>(undefined);
 
@@ -82,12 +46,6 @@ const Results = ({ type }: ResultsProps) => {
   const [loading, setLoading] = useState(true);
   const [apiError, setApiError] = useState(false);
   const [apiResults, setApiResults] = useState<EligibilityResults | undefined>();
-
-  const [value, setValue] = useState(0);
-
-  const handleChange = (event: React.SyntheticEvent, newValue: number) => {
-    setValue(newValue);
-  };
 
   const fetchResults = async () => {
     try {
@@ -142,27 +100,7 @@ const Results = ({ type }: ResultsProps) => {
         }}
       >
         <ResultsHeader type={type} />
-        <Box sx={{ width: '100%' }}>
-          <Box className="results-tab-container">
-            <Tabs
-              value={value}
-              onChange={handleChange}
-              aria-label="basic tabs example"
-              variant="fullWidth"
-              indicatorColor="secondary"
-            >
-              <Tab label={`LONG-TERM BENEFITS (${programs.length})`} {...a11yProps(0)} />
-              <Tab label={`NEAR-TERM RESOURCES (${needs.length})`} {...a11yProps(1)} />
-            </Tabs>
-          </Box>
-          <CustomTabPanel value={value} index={0}>
-            <ResultsTabs currentTab={type} />
-            {type === 'need' ? <Needs /> : <Programs />}
-          </CustomTabPanel>
-          <CustomTabPanel value={value} index={1}>
-            <Needs />
-          </CustomTabPanel>
-        </Box>
+        <ResultsTabs currentTab={type} />
         <MoreHelp />
       </ResultsContext.Provider>
     );
